update syntax, demo
