+++ CMakeLists.txt
@@ -466,48 +466,35 @@ endif()
 
 # =============================================== Demo/test ===============================================
 
+macro(add_test_dependencies exec_name)
+  target_compile_features(${exec_name} PRIVATE cxx_std_11)
+  target_link_libraries(${exec_name} PRIVATE ${target_name})
+  # Clang has been fast to adopt dwarf 5, other tools (e.g. addr2line from binutils) have not
+  check_cxx_compiler_flag("-gdwarf-4" HAS_DWARF4)
+  if(HAS_DWARF4)
+    target_compile_options(${exec_name} PRIVATE "$<$<CONFIG:Debug>:-gdwarf-4>")
+  endif()
+  # TODO: add debug info for mingw clang?
+  if(CPPTRACE_BUILD_TEST_RDYNAMIC)
+    set_property(TARGET ${exec_name} PROPERTY ENABLE_EXPORTS ON)
+  endif()
+  if(APPLE) # TODO: Temporary
+    add_custom_command(
+      TARGET ${exec_name}
+      POST_BUILD
+      COMMAND dsymutil $<TARGET_FILE:${exec_name}>
+    )
+  endif()
+endmacro()
+
 if(CPPTRACE_BUILD_TESTING)
   add_executable(test test/test.cpp)
-  target_compile_features(test PRIVATE cxx_std_11)
-  target_link_libraries(test PRIVATE ${target_name})
-  # Clang has been fast to adopt dwarf 5, other tools (e.g. addr2line from binutils) have not
-  check_cxx_compiler_flag("-gdwarf-4" HAS_DWARF4)
-  if(HAS_DWARF4)
-    target_compile_options(test PRIVATE "$<$<CONFIG:Debug>:-gdwarf-4>")
-  endif()
-  if(CPPTRACE_BUILD_TEST_RDYNAMIC)
-    set_property(TARGET test PROPERTY ENABLE_EXPORTS ON)
-  endif()
-  if(APPLE) # TODO: Temporary
-    add_custom_command(
-      TARGET test
-      POST_BUILD
-      COMMAND dsymutil $<TARGET_FILE:test>
-    )
-  endif()
-
   add_executable(demo test/demo.cpp)
-  target_compile_features(demo PRIVATE cxx_std_11)
-  target_link_libraries(demo PRIVATE ${target_name})
-  # Clang has been fast to adopt dwarf 5, other tools (e.g. addr2line from binutils) have not
-  check_cxx_compiler_flag("-gdwarf-4" HAS_DWARF4)
-  if(HAS_DWARF4)
-    target_compile_options(demo PRIVATE "$<$<CONFIG:Debug>:-gdwarf-4>")
-  endif()
-  if(WIN32 AND CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
-    target_compile_options(demo PUBLIC -g -gcodeview)
-    target_link_options(demo PUBLIC -fuse-ld=lld -g -Wl,--pdb=)
-  endif()
-  if(CPPTRACE_BUILD_TEST_RDYNAMIC)
-    set_property(TARGET demo PROPERTY ENABLE_EXPORTS ON)
-  endif()
-  if(APPLE) # TODO: Temporary
-    add_custom_command(
-      TARGET demo
-      POST_BUILD
-      COMMAND dsymutil $<TARGET_FILE:demo>
-    )
-  endif()
+  add_executable(c_demo test/ctrace_demo.cpp)
+  
+  add_test_dependencies(test)
+  add_test_dependencies(demo)
+  add_test_dependencies(c_demo)
 
   if(UNIX)
     add_executable(signal_demo test/signal_demo.cpp)
